chore(src): reorganize Typography elements
