+++ src/helix-ui/elements/hx-error/_shadow.less
@@ -1,4 +1,4 @@
-@import './HXElement';
+@import '../HXElement';
 
 #hxError {
   display: inline-flex;
